refactor extra person events into mock database
